# src/main/java/br/com/gustavo/votacao/entity/Associado.java
package br.com.gustavo.votacao.entity;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@Builder
@Data
@NoArgsConstructor
public class Associado {

    @JsonProperty("cpf")
    private String cpf;
}